JBIDE-12029 - No type for PathParam parameters in JAX-RS REST Web Services explorer
Changing the way the endpoint's URI Template are built from the @Path, @QueryParam
and @MatrixParam annotations on the Java types, fields and methods.
Applied changes to the UI side with new URLTemplateParameter and URLTemplateParser
classes that deal with parsing, validating and replacing the params in the 'stashedURL'

Added JUnit tests.

Changed the Bundle-RequiredExecutionEnvironment to "JavaSE-1.6" in MANIFEST.MF of
org.jboss.tools.ws.ui.test plugin to allow for Java6 @Override syntax on interface
methods implementation.