JIBDE-12860 - JAX-RS validation problems are not linked
Fixed the problem where some markers would not be removed from the resource to be validated between
two validation calls, which meant that duplicate markers for the same problem would exist on each
resource.

Also refactored the search/element retrieval: the JaxrsElementIndexationDelegate now not only
deals with the Lucene indexation and searches, but also uses the metamodel to return List of
JAX-RS elements, Endpoints or workspace Resources.
Renamed some marker deletion/undexing methods to use the "remove" verb (eg: "removeMarker()") and be consistent with the
existing "remove()" methods that remove the elements themselves.
Added JUnit tests to covers the fixes.

This patch also fixes:
- JBIDE-15662: Warning Multiple JAX-RS Activators doesn't disappear
- JBIDE-15664: Application path configured in web.xml causes warning