JBIDE-15908 - Problem decorator is not removed after error was fixed
Bug was that the problemLevel on each JAX-RS Resource Method would not
be reset at the beginning of the validation, and during validation, the
highest problem level is kept. Thus, during the next validation (fixing
the problem), the problem level would remain at its highest (previous)
value.

Added tests to verify the correct behaviour on all validators.